// src/services/api/generated/models/test-pcstatus-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `ONLINE` - Online * `OFFLINE` - Offline * `MAINTENANCE` - Maintenance * `INITIALIZING` - Initializing
 */

export const TestPCStatusEnum = {
    ONLINE: 'ONLINE',
    OFFLINE: 'OFFLINE',
    MAINTENANCE: 'MAINTENANCE',
    INITIALIZING: 'INITIALIZING'
} as const;

export type TestPCStatusEnum = typeof TestPCStatusEnum[keyof typeof TestPCStatusEnum];